Replace rhea with own size and position calculations, builds, not tested

+++ src/ttauri/thread_win32.cpp
@@ -33,7 +33,7 @@ void run_from_main_loop(std::function<void()> f)
 
 void wait_on(std::atomic<uint32_t> &value, uint32_t expected, hires_utc_clock::duration timeout) noexcept
 {
-    DWORD timeout_ms = timeout == hires_utc_clock::duration::max() ? INFINITE : timeout / 1ms;
+    DWORD timeout_ms = timeout == hires_utc_clock::duration::max() ? INFINITE : numeric_cast<DWORD>(timeout / 1ms);
     if (!WaitOnAddress(&value, &expected, sizeof (value), timeout_ms)) {
         if (GetLastError() != ERROR_TIMEOUT) {
             LOG_FATAL("Could not wait on address {}", getLastErrorMessage());
